+++ src/sealed_float.rs
@@ -43,6 +43,9 @@ pub trait SealedFloat: Copy {
     fn infinity(neg: bool) -> Self;
     fn from_parts(neg: bool, exp: i32, mant: Self::Bits) -> Self;
     fn parts(self) -> (bool, i32, Self::Bits);
+
+    fn from_neg_abs(neg: bool, abs: u128, frac_bits: u32, int_bits: u32) -> Self;
+    fn to_neg_abs(self, frac_bits: u32, int_bits: u32) -> Option<(bool, u128)>;
 }
 
 macro_rules! sealed_float {
@@ -104,6 +107,122 @@ macro_rules! sealed_float {
                 let mant = bits & mant_mask;
 
                 (neg, exp, mant)
+            }
+
+            fn from_neg_abs(neg: bool, abs: u128, frac_bits: u32, int_bits: u32) -> $Float {
+                let prec = Self::prec();
+                let exp_min = Self::exp_min();
+                let exp_max = Self::exp_max();
+                let fix_bits = frac_bits + int_bits;
+
+                let extra_zeros = 128 - fix_bits;
+                let leading_zeros = abs.leading_zeros() - extra_zeros;
+                let signif_bits = fix_bits - leading_zeros;
+                if signif_bits == 0 {
+                    return Self::zero(neg);
+                }
+                // remove leading zeros and implicit one
+                let mut mantissa = abs << leading_zeros << 1;
+                let exponent = int_bits as i32 - 1 - leading_zeros as i32;
+                let biased_exponent = if exponent > exp_max {
+                    return Self::infinity(neg);
+                } else if exponent < exp_min {
+                    let lost_prec = exp_min - exponent;
+                    if lost_prec as u32 >= (int_bits + frac_bits) {
+                        mantissa = 0;
+                    } else {
+                        // reinsert implicit one
+                        mantissa = (mantissa >> 1) | !(!0 >> 1);
+                        mantissa >>= lost_prec - 1;
+                    }
+                    0
+                } else {
+                    (exponent + exp_max) as Self::Bits
+                };
+                // check for rounding
+                let round_up = (fix_bits >= prec) && {
+                    let shift = prec - 1;
+                    let mid_bit = !(!0 >> 1) >> (shift + extra_zeros);
+                    let lower_bits = mid_bit - 1;
+                    if mantissa & mid_bit == 0 {
+                        false
+                    } else if mantissa & lower_bits != 0 {
+                        true
+                    } else {
+                        // round to even
+                        mantissa & (mid_bit << 1) != 0
+                    }
+                };
+                let bits_sign = if neg { !(!0 >> 1) } else { 0 };
+                let bits_exp = biased_exponent << (prec - 1);
+                let bits_mantissa = (if fix_bits >= prec - 1 {
+                    (mantissa >> (fix_bits - (prec - 1))) as Self::Bits
+                } else {
+                    (mantissa as Self::Bits) << (prec - 1 - fix_bits)
+                }) & !(!0 << (prec - 1));
+                let mut bits_exp_mantissa = bits_exp | bits_mantissa;
+                if round_up {
+                    bits_exp_mantissa += 1;
+                }
+                Self::from_bits(bits_sign | bits_exp_mantissa)
+            }
+
+            fn to_neg_abs(self, frac_bits: u32, int_bits: u32) -> Option<(bool, u128)> {
+                let float_bits = Self::Bits::nbits() as i32;
+                let prec = Self::prec() as i32;
+                let exp_min = Self::exp_min();
+                let exp_max = Self::exp_max();
+                let fix_bits = (frac_bits + int_bits) as i32;
+
+                let (neg, exp, mut mantissa) = self.parts();
+                if exp > exp_max {
+                    return None;
+                }
+                // if not subnormal, add implicit bit
+                if exp >= exp_min {
+                    mantissa |= 1 << (prec - 1);
+                }
+                if mantissa == 0 {
+                    return Some((neg, 0));
+                }
+                let leading_zeros = mantissa.leading_zeros();
+                mantissa <<= leading_zeros;
+                let mut need_to_shr = -exp + leading_zeros as i32 + prec - 1 - frac_bits as i32;
+                let rounding_bit = need_to_shr - 1;
+                if 0 <= rounding_bit
+                    && rounding_bit < float_bits
+                    && mantissa & (1 << rounding_bit) != 0
+                {
+                    // Rounding bit is one.
+                    // If any lower bit is one, round up.
+                    // If bit exactly above rounding is one, round up (tie to even).
+                    let round_up = (rounding_bit > 0 && mantissa & ((1 << rounding_bit) - 1) != 0)
+                        || (rounding_bit + 1 < float_bits
+                            && mantissa & (1 << (rounding_bit + 1)) != 0);
+                    if round_up {
+                        mantissa = match mantissa.overflowing_add(1 << rounding_bit) {
+                            (m, false) => m,
+                            (m, true) => {
+                                need_to_shr -= 1;
+                                (m >> 1) | (1 << (float_bits - 1))
+                            }
+                        };
+                    }
+                }
+                // now rounding is done, we can truncate extra right bits
+                if float_bits - need_to_shr > fix_bits {
+                    return None;
+                }
+                let abs = if need_to_shr == 0 {
+                    u128::from(mantissa)
+                } else if need_to_shr < 0 {
+                    u128::from(mantissa) << -need_to_shr
+                } else if need_to_shr < float_bits {
+                    u128::from(mantissa >> need_to_shr)
+                } else {
+                    0
+                };
+                Some((neg, abs))
             }
         }
     };